Simplified service deployment on EventBus
In our CQRS context verticles will communicate only with each other
through the EventBus. Their communication is not meant to be exposed to
the outside. So we can skip Discovery service for now since it makes
deploying and consuming EBServices much simpler.

Test are also improved to initialize verticles in the correct order and
inject configuration parameters in order to know on which port to call
the facade.