Clarified application launch
See discussion at
https://groups.google.com/d/msg/vertx/dnvWA7O67dM/ltsG6F2zBQAJ for
detailed investigation.

The final implementation will consist in :
- a default Verticle that will handle the « all-in-one » default
deployment
- using the default Vert.X « run » command to launch another single
Verticle.

If the need arises, the default verticle could be improved with a
configuration file specifying which verticles should be deployed and
how many, for example to deploy Command and EventStore but not Query.

I updated (and tested) the bash script to reflect the command-line
solution I found to launch a specific verticle in a background process.
This file is no longer used but I keep it for further reference.