First implementation of Indicator Groups in Scales
- The tirival initial structure of indicators has been replaced with a
hierarchical structure
- The IndicatorGroupVO is created as a ValueObject in the
ScaleAggregate. Logic for handling Indicators has been mostly migrated
into this class, as a Scale now have a single root IndicatorGroupVO and
encapsulates some of its API.
- The ScaleElement and IndicatorElement classes are package-level
classes used only for internal implementation. They are not part of the
conceptual Aggregate itself.

Note: the Indicator ordering scenarios are failing for now, as the
implementation is not complete.